+++ context/webpack.config.js
@@ -7,6 +7,7 @@ const config = {
     filename: 'bundle.js',
     publicPath: resolve('../app/static/public'),
   },
+  devtool: 'nosources-source-map',
   resolve: {
     extensions: ['.js', '.jsx', '.css', '.woff', '.woff2', '.svg'],
   },
@@ -22,6 +23,11 @@ const config = {
           emitWarning: true,
           configFile: './.eslintrc.yml',
         },
+      },
+      {
+        test: /\.js$/,
+        use: ['source-map-loader'],
+        enforce: 'pre',
       },
       {
         test: /\.jsx?/,
@@ -40,6 +46,14 @@ const config = {
           {
             loader: 'html-loader',
           },
+        ],
+      },
+      {
+        test: /\.s[ac]ss$/i,
+        use: [
+          'style-loader',
+          'css-loader',
+          'sass-loader',
         ],
       },
       {